orchestration/pyproject.toml
[project]
name = "pixeltable-orchestration"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pixeltable>=0.5.19",
    "pillow>=10.0",
    "psycopg[binary]>=3.1",
    "sqlalchemy>=2.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
